requirements.txt
pbr>=0.5.21,<1.0
SQLObject>=1.3.2
xworkflows
rarfile>=2.6
shutilwhich
lockfile
BitTorrent-bencode
stevedore>=0.10
six>=1.4.1
oslo.config>=1.2.0
blinker>=1.3
ordereddict>=1.1

